Extract service config provider out of configuration
